app/requirements.txt
alembic==1.0.3
aniso8601==4.0.1
bcrypt==3.1.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==1.1.15
cymem==1.31.2
cytoolz==0.9.0.1
dill==0.2.8.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.3.0
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
idna==2.7
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.3.1
murmurhash==0.28.0
numpy==1.15.1
plac==0.9.6
preshed==1.0.1
pycparser==2.19
PyJWT==1.6.4
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2018.7
PyYAML==3.13
regex==2017.4.5
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.14
swagger-spec-validator==2.4.1
thinc==6.10.3
toolz==0.9.0
tqdm==4.25.0
ujson==1.35
urllib3==1.23
virtualenv==16.1.0
Werkzeug==0.14.1
wrapt==1.10.11
